Attr 第二个参数修改为 enum，同时支持 isset/get/update 3种语义

+++ tests/src/base.cpp
@@ -74,7 +74,7 @@ TEST(base, echo) {
     var b = 2026;
     echo(b);
     echo("\n");
-    echo(1987L);
+    echo((Int) 1987L);
     echo("\n");
     echo(3.1415926);
     echo("\n");
